added app and db files
